// resources/js/api/generated/models/composeService.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { Container } from './container';

export interface ComposeService {
  containers: Container[];
  image?: string;
  name: string;
  ports?: string[];
}